+++ src/test/java/api/loans/CheckOutByBarcodeTests.java
@@ -56,10 +56,6 @@ import java.util.UUID;
 import java.util.concurrent.TimeUnit;
 
 import org.awaitility.Awaitility;
-import org.folio.circulation.domain.override.BlockOverrides;
-import org.folio.circulation.domain.override.ItemLimitBlockOverride;
-import org.folio.circulation.domain.override.ItemNotLoanableBlockOverride;
-import org.folio.circulation.domain.override.PatronBlockOverride;
 import org.folio.circulation.domain.policy.Period;
 import org.folio.circulation.support.http.client.Response;
 import org.joda.time.DateTime;
@@ -68,10 +64,12 @@ import org.joda.time.Seconds;
 import org.junit.Test;
 
 import api.support.APITests;
+import api.support.builders.CheckOutBlockOverrides;
 import api.support.builders.CheckOutByBarcodeRequestBuilder;
 import api.support.builders.FixedDueDateSchedule;
 import api.support.builders.FixedDueDateSchedulesBuilder;
 import api.support.builders.ItemBuilder;
+import api.support.builders.ItemNotLoanableBlockOverrideBuilder;
 import api.support.builders.LoanBuilder;
 import api.support.builders.LoanPolicyBuilder;
 import api.support.builders.NoticePolicyBuilder;
@@ -101,6 +99,7 @@ public class CheckOutByBarcodeTests extends APITests {
     "Insufficient override permissions";
   private static final String TEST_COMMENT = "Some comment";
   private static final String CHECKED_OUT_THROUGH_OVERRIDE = "checkedOutThroughOverride";
+  private static final String PATRON_WAS_BLOCKED_MESSAGE = "Patron blocked from borrowing";
 
   @Test
   public void canCheckOutUsingItemAndUserBarcode() {
@@ -1416,6 +1415,105 @@ public class CheckOutByBarcodeTests extends APITests {
   }
 
   @Test
+  public void cannotOverrideItemNotLoanableBlockWhenDueDateIsBeforeLoanDate() {
+    final OkapiHeaders okapiHeaders = buildOkapiHeadersWithPermissions(
+      OVERRIDE_ITEM_NOT_LOANABLE_BLOCK_PERMISSION);
+
+    DateTime invalidDueDate = TEST_LOAN_DATE.minusDays(2);
+
+    setNotLoanablePolicy();
+    Response response = checkOutFixture.attemptCheckOutByBarcode(
+      new CheckOutByBarcodeRequestBuilder()
+        .forItem(itemsFixture.basedUponSmallAngryPlanet())
+        .to(usersFixture.steve())
+        .at(UUID.randomUUID())
+        .on(TEST_LOAN_DATE)
+        .withOverrideBlocks(new CheckOutBlockOverrides()
+          .withItemNotLoanableBlockOverride(new ItemNotLoanableBlockOverrideBuilder()
+            .withDueDate(invalidDueDate)
+            .create())
+          .withComment(TEST_COMMENT)
+          .create()),
+      okapiHeaders);
+
+    assertThat(response.getJson(), hasErrorWith(allOf(
+      hasMessage("Due date should be later than loan date"),
+      hasParameter("dueDate", invalidDueDate.toString()))));
+  }
+
+  @Test
+  public void cannotOverrideItemNotLoanableBlockWhenDueDateIsTheSameAsLoanDate() {
+    final OkapiHeaders okapiHeaders = buildOkapiHeadersWithPermissions(
+      OVERRIDE_ITEM_NOT_LOANABLE_BLOCK_PERMISSION);
+
+    setNotLoanablePolicy();
+    Response response = checkOutFixture.attemptCheckOutByBarcode(
+      new CheckOutByBarcodeRequestBuilder()
+        .forItem(itemsFixture.basedUponSmallAngryPlanet())
+        .to(usersFixture.steve())
+        .at(UUID.randomUUID())
+        .on(TEST_LOAN_DATE)
+        .withOverrideBlocks(new CheckOutBlockOverrides()
+          .withItemNotLoanableBlockOverride(new ItemNotLoanableBlockOverrideBuilder()
+            .withDueDate(TEST_LOAN_DATE)
+            .create())
+          .withComment(TEST_COMMENT)
+          .create()),
+      okapiHeaders);
+
+    assertThat(response.getJson(), hasErrorWith(allOf(
+      hasMessage("Due date should be later than loan date"),
+      hasParameter("dueDate", TEST_LOAN_DATE.toString()))));
+  }
+
+  @Test
+  public void cannotOverrideBlockWhenCommentIsNotPresent() {
+    final OkapiHeaders okapiHeaders = buildOkapiHeadersWithPermissions(
+      OVERRIDE_ITEM_NOT_LOANABLE_BLOCK_PERMISSION);
+
+    setNotLoanablePolicy();
+    Response response = checkOutFixture.attemptCheckOutByBarcode(
+      new CheckOutByBarcodeRequestBuilder()
+        .forItem(itemsFixture.basedUponSmallAngryPlanet())
+        .to(usersFixture.steve())
+        .at(UUID.randomUUID())
+        .on(TEST_LOAN_DATE)
+        .withOverrideBlocks(new CheckOutBlockOverrides()
+          .withItemNotLoanableBlockOverride(new ItemNotLoanableBlockOverrideBuilder()
+            .withDueDate(TEST_DUE_DATE)
+            .create())
+          .create()),
+      okapiHeaders);
+
+    assertThat(response.getJson(), hasErrorWith(allOf(
+      hasMessage("Override should be performed with the comment specified"),
+      hasParameter("comment", null))));
+  }
+
+  @Test
+  public void cannotOverrideItemNotLoanableBlockWhenDueDateIsNotPresent() {
+    final OkapiHeaders okapiHeaders = buildOkapiHeadersWithPermissions(
+      OVERRIDE_ITEM_NOT_LOANABLE_BLOCK_PERMISSION);
+
+    setNotLoanablePolicy();
+    Response response = checkOutFixture.attemptCheckOutByBarcode(
+      new CheckOutByBarcodeRequestBuilder()
+        .forItem(itemsFixture.basedUponSmallAngryPlanet())
+        .to(usersFixture.steve())
+        .at(UUID.randomUUID())
+        .on(TEST_LOAN_DATE)
+        .withOverrideBlocks(new CheckOutBlockOverrides()
+          .withItemNotLoanableBlockOverride(new JsonObject())
+          .withComment(TEST_COMMENT)
+          .create()),
+      okapiHeaders);
+
+    assertThat(response.getJson(), hasErrorWith(allOf(
+      hasMessage("Override should be performed with due date specified"),
+      hasParameter("dueDate", null))));
+  }
+
+  @Test
   public void cannotOverrideItemNotLoanableBlockWhenUserDoesNotHavePermissions() {
     setNotLoanablePolicy();
     Response response = checkOutFixture.attemptCheckOutByBarcode(
@@ -1424,9 +1522,13 @@ public class CheckOutByBarcodeTests extends APITests {
         .to(usersFixture.steve())
         .at(UUID.randomUUID())
         .on(TEST_LOAN_DATE)
-        .withOverrideBlocks(
-          new BlockOverrides(new ItemNotLoanableBlockOverride(true, TEST_DUE_DATE),
-            null, new ItemLimitBlockOverride(false), TEST_COMMENT)));
+        .withOverrideBlocks(new CheckOutBlockOverrides()
+          .withItemNotLoanableBlockOverride(new ItemNotLoanableBlockOverrideBuilder()
+            .withDueDate(TEST_LOAN_DATE)
+            .create())
+          .withComment(TEST_COMMENT)
+          .create()));
+
     assertThat(response.getJson(), hasErrorWith(hasMessage(INSUFFICIENT_OVERRIDE_PERMISSIONS)));
     assertThat(getMissingPermissions(response), hasSize(1));
     assertThat(getMissingPermissions(response),
@@ -1444,8 +1546,13 @@ public class CheckOutByBarcodeTests extends APITests {
         .forItem(itemsFixture.basedUponSmallAngryPlanet())
         .to(usersFixture.steve())
         .at(UUID.randomUUID())
-        .on(TEST_LOAN_DATE).withOverrideBlocks(new BlockOverrides(
-        new ItemNotLoanableBlockOverride(true, TEST_DUE_DATE), null, null, TEST_COMMENT)),
+        .on(TEST_LOAN_DATE)
+        .withOverrideBlocks(new CheckOutBlockOverrides()
+          .withItemNotLoanableBlockOverride(new ItemNotLoanableBlockOverrideBuilder()
+            .withDueDate(TEST_DUE_DATE)
+            .create())
+          .withComment(TEST_COMMENT)
+          .create()),
       okapiHeaders);
 
     assertThat(response.getJson(), hasErrorWith(hasMessage(INSUFFICIENT_OVERRIDE_PERMISSIONS)));
@@ -1462,9 +1569,14 @@ public class CheckOutByBarcodeTests extends APITests {
         .forItem(itemsFixture.basedUponSmallAngryPlanet())
         .to(usersFixture.steve())
         .at(UUID.randomUUID())
-        .on(TEST_LOAN_DATE).withOverrideBlocks(new BlockOverrides(
-        new ItemNotLoanableBlockOverride(true, TEST_DUE_DATE), new PatronBlockOverride(true),
-        null, TEST_COMMENT)));
+        .on(TEST_LOAN_DATE)
+        .withOverrideBlocks(new CheckOutBlockOverrides()
+          .withItemNotLoanableBlockOverride(new ItemNotLoanableBlockOverrideBuilder()
+            .withDueDate(TEST_DUE_DATE)
+            .create())
+          .withPatronBlockOverride(new JsonObject())
+          .withComment(TEST_COMMENT)
+          .create()));
 
     assertThat(response.getJson(), hasErrorWith(hasMessage(INSUFFICIENT_OVERRIDE_PERMISSIONS)));
     assertThat(getMissingPermissions(response), hasSize(2));
@@ -1479,8 +1591,10 @@ public class CheckOutByBarcodeTests extends APITests {
         .forItem(itemsFixture.basedUponNod())
         .to(usersFixture.steve())
         .at(UUID.randomUUID())
-        .withOverrideBlocks(new BlockOverrides(
-          null, null, new ItemLimitBlockOverride(true), TEST_COMMENT)));
+        .withOverrideBlocks(new CheckOutBlockOverrides()
+          .withItemLimitBlockOverride(new JsonObject())
+          .withComment(TEST_COMMENT)
+          .create()));
 
     assertThat(response.getJson(), hasErrorWith(hasMessage(INSUFFICIENT_OVERRIDE_PERMISSIONS)));
     assertThat(getMissingPermissions(response), hasSize(1));
@@ -1494,8 +1608,10 @@ public class CheckOutByBarcodeTests extends APITests {
         .forItem(itemsFixture.basedUponNod())
         .to(usersFixture.steve())
         .at(UUID.randomUUID())
-        .withOverrideBlocks(new BlockOverrides(
-          null, new PatronBlockOverride(true), null, TEST_COMMENT)));
+        .withOverrideBlocks(new CheckOutBlockOverrides()
+          .withPatronBlockOverride(new JsonObject())
+          .withComment(TEST_COMMENT)
+          .create()));
 
     assertThat(response.getJson(), hasErrorWith(hasMessage(INSUFFICIENT_OVERRIDE_PERMISSIONS)));
     assertThat(getMissingPermissions(response), hasSize(1));
@@ -1512,8 +1628,10 @@ public class CheckOutByBarcodeTests extends APITests {
         .forItem(itemsFixture.basedUponNod())
         .to(usersFixture.steve())
         .at(UUID.randomUUID())
-        .withOverrideBlocks(new BlockOverrides(
-          null, new PatronBlockOverride(true), null, TEST_COMMENT)),
+        .withOverrideBlocks(new CheckOutBlockOverrides()
+          .withPatronBlockOverride(new JsonObject())
+          .withComment(TEST_COMMENT)
+          .create()),
       okapiHeaders);
 
     assertThat(response.getJson(), hasErrorWith(hasMessage(INSUFFICIENT_OVERRIDE_PERMISSIONS)));
@@ -1544,8 +1662,10 @@ public class CheckOutByBarcodeTests extends APITests {
         .forItem(secondBookTypeItem)
         .to(steve)
         .at(UUID.randomUUID())
-        .withOverrideBlocks(new BlockOverrides(
-          null, null, new ItemLimitBlockOverride(true), TEST_COMMENT)),
+        .withOverrideBlocks(new CheckOutBlockOverrides()
+          .withItemLimitBlockOverride(new JsonObject())
+          .withComment(TEST_COMMENT)
+          .create()),
       okapiHeaders).getJson();
 
     secondBookTypeItem = itemsClient.get(secondBookTypeItem);
@@ -1567,7 +1687,9 @@ public class CheckOutByBarcodeTests extends APITests {
         .to(usersFixture.steve())
         .at(UUID.randomUUID())
         .on(TEST_LOAN_DATE)
-        .withOverrideBlocks(new BlockOverrides(null, null, null, TEST_COMMENT)),
+        .withOverrideBlocks(new CheckOutBlockOverrides()
+          .withComment(TEST_COMMENT)
+          .create()),
       okapiHeaders);
 
     assertThat(responseBlocked.getJson(), hasErrorWith(allOf(hasMessage("Item is not loanable"),
@@ -1579,15 +1701,79 @@ public class CheckOutByBarcodeTests extends APITests {
         .to(usersFixture.steve())
         .at(UUID.randomUUID())
         .on(TEST_LOAN_DATE)
-        .withOverrideBlocks(new BlockOverrides(
-          new ItemNotLoanableBlockOverride(true, new DateTime(TEST_DUE_DATE)),
-          null, null, TEST_COMMENT)),
+        .withOverrideBlocks(new CheckOutBlockOverrides()
+          .withItemNotLoanableBlockOverride(new ItemNotLoanableBlockOverrideBuilder()
+            .withDueDate(TEST_DUE_DATE)
+            .create())
+          .withComment(TEST_COMMENT)
+          .create()),
       okapiHeaders).getJson();
 
     item = itemsClient.get(item);
     assertThat(item, hasItemStatus(CHECKED_OUT));
     assertThat(loan.getString("actionComment"), is(TEST_COMMENT));
     assertThat(loan.getString("action"), is(CHECKED_OUT_THROUGH_OVERRIDE));
+  }
+
+  @Test
+  public void canOverrideCheckOutWhenItemIsLoanableAndOverrideIsRequested() {
+    final OkapiHeaders okapiHeaders = buildOkapiHeadersWithPermissions(
+      OVERRIDE_ITEM_NOT_LOANABLE_BLOCK_PERMISSION);
+
+    IndividualResource item = itemsFixture.basedUponSmallAngryPlanet();
+
+    JsonObject loan = checkOutFixture.checkOutByBarcode(
+      new CheckOutByBarcodeRequestBuilder()
+        .forItem(item)
+        .to(usersFixture.steve())
+        .at(UUID.randomUUID())
+        .on(TEST_LOAN_DATE)
+        .withOverrideBlocks(new CheckOutBlockOverrides()
+          .withItemNotLoanableBlockOverride(new ItemNotLoanableBlockOverrideBuilder()
+            .withDueDate(TEST_DUE_DATE)
+            .create())
+          .withComment(TEST_COMMENT)
+          .create()),
+      okapiHeaders).getJson();
+
+    item = itemsClient.get(item);
+    assertThat(item, hasItemStatus(CHECKED_OUT));
+    assertThat(loan.getString("actionComment"), is(TEST_COMMENT));
+    assertThat(loan.getString("action"), is(CHECKED_OUT_THROUGH_OVERRIDE));
+  }
+
+  @Test
+  public void canCreateRecallRequestAfterOverriddenCheckout() {
+    final OkapiHeaders okapiHeaders = buildOkapiHeadersWithPermissions(
+      OVERRIDE_ITEM_NOT_LOANABLE_BLOCK_PERMISSION);
+    IndividualResource item = itemsFixture.basedUponSmallAngryPlanet();
+
+    IndividualResource steve = usersFixture.steve();
+    IndividualResource charlotte = usersFixture.charlotte();
+
+    setNotLoanablePolicy();
+
+    checkOutFixture.checkOutByBarcode(
+      new CheckOutByBarcodeRequestBuilder()
+        .forItem(item)
+        .to(steve)
+        .at(servicePointsFixture.cd1())
+        .on(TEST_LOAN_DATE)
+        .withOverrideBlocks(new CheckOutBlockOverrides()
+          .withItemNotLoanableBlockOverride(new ItemNotLoanableBlockOverrideBuilder()
+            .withDueDate(TEST_DUE_DATE)
+            .create())
+          .withComment(TEST_COMMENT)
+          .create()),
+      okapiHeaders).getJson();
+
+    final Response placeRequestResponse = requestsFixture.attemptPlace(new RequestBuilder()
+      .recall()
+      .forItem(item)
+      .by(charlotte)
+      .fulfilToHoldShelf(servicePointsFixture.cd1()));
+
+    assertThat(placeRequestResponse.getStatusCode(), is(201));
   }
 
   @Test
@@ -1612,8 +1798,10 @@ public class CheckOutByBarcodeTests extends APITests {
         .to(steve)
         .at(UUID.randomUUID())
         .on(TEST_LOAN_DATE)
-        .withOverrideBlocks(new BlockOverrides(
-          null, new PatronBlockOverride(true), null, TEST_COMMENT)),
+        .withOverrideBlocks(new CheckOutBlockOverrides()
+          .withPatronBlockOverride(new JsonObject())
+          .withComment(TEST_COMMENT)
+          .create()),
       okapiHeaders).getJson();
 
     item = itemsClient.get(item);
@@ -1636,8 +1824,10 @@ public class CheckOutByBarcodeTests extends APITests {
         .to(steve)
         .at(UUID.randomUUID())
         .on(TEST_LOAN_DATE)
-        .withOverrideBlocks(new BlockOverrides(null, new PatronBlockOverride(true), null,
-          TEST_COMMENT)));
+        .withOverrideBlocks(new CheckOutBlockOverrides()
+          .withPatronBlockOverride(new JsonObject())
+          .withComment(TEST_COMMENT)
+          .create()));
 
     assertThat(response.getJson(), hasErrorWith(hasMessage(INSUFFICIENT_OVERRIDE_PERMISSIONS)));
     assertThat(response.getJson(), hasErrorWith(hasMessage("Item is already checked out")));
@@ -1645,6 +1835,75 @@ public class CheckOutByBarcodeTests extends APITests {
       "Cannot check out item that already has an open loan")));
     assertThat(getMissingPermissions(response), hasSize(1));
     assertThat(getMissingPermissions(response), hasItem(OVERRIDE_PATRON_BLOCK_PERMISSION));
+  }
+
+  @Test
+  public void canOverrideManualPatronBlockWhenBlockIsPresent() {
+    IndividualResource item = itemsFixture.basedUponSmallAngryPlanet();
+    final IndividualResource steve = usersFixture.steve();
+    userManualBlocksFixture.createManualPatronBlockForUser(steve.getId());
+
+    final Response response = checkOutFixture.attemptCheckOutByBarcode(item, steve);
+
+    assertThat(response, hasStatus(HTTP_UNPROCESSABLE_ENTITY));
+    assertThat(response.getJson(),
+      hasErrorWith(hasMessage(PATRON_WAS_BLOCKED_MESSAGE)));
+
+    final OkapiHeaders okapiHeaders = buildOkapiHeadersWithPermissions(
+      OVERRIDE_PATRON_BLOCK_PERMISSION);
+    JsonObject loan = checkOutFixture.checkOutByBarcode(
+      new CheckOutByBarcodeRequestBuilder()
+        .forItem(item)
+        .to(steve)
+        .at(UUID.randomUUID())
+        .on(TEST_LOAN_DATE)
+        .withOverrideBlocks(new CheckOutBlockOverrides()
+          .withPatronBlockOverride(new JsonObject())
+          .withComment(TEST_COMMENT)
+          .create()),
+      okapiHeaders).getJson();
+
+    item = itemsClient.get(item);
+    assertThat(item, hasItemStatus(CHECKED_OUT));
+    assertThat(loan.getString("actionComment"), is(TEST_COMMENT));
+    assertThat(loan.getString("action"), is(CHECKED_OUT_THROUGH_OVERRIDE));
+  }
+
+  @Test
+  public void canOverrideManualAndAutomationPatronBlocksWhenBlocksArePresent() {
+    IndividualResource item = itemsFixture.basedUponSmallAngryPlanet();
+    final IndividualResource steve = usersFixture.steve();
+    userManualBlocksFixture.createManualPatronBlockForUser(steve.getId());
+    automatedPatronBlocksFixture.blockAction(steve.getId().toString(), true, false, false);
+
+    final Response response = checkOutFixture.attemptCheckOutByBarcode(item, steve);
+
+    assertThat(response, hasStatus(HTTP_UNPROCESSABLE_ENTITY));
+    assertThat(response.getJson(),
+      hasErrorWith(hasMessage(PATRON_WAS_BLOCKED_MESSAGE)));
+    assertThat(response.getJson(),
+      hasErrorWith(hasMessage(MAX_NUMBER_OF_ITEMS_CHARGED_OUT_MESSAGE)));
+    assertThat(response.getJson(),
+      hasErrorWith(hasMessage(MAX_OUTSTANDING_FEE_FINE_BALANCE_MESSAGE)));
+
+    final OkapiHeaders okapiHeaders = buildOkapiHeadersWithPermissions(
+      OVERRIDE_PATRON_BLOCK_PERMISSION);
+    JsonObject loan = checkOutFixture.checkOutByBarcode(
+      new CheckOutByBarcodeRequestBuilder()
+        .forItem(item)
+        .to(steve)
+        .at(UUID.randomUUID())
+        .on(TEST_LOAN_DATE)
+        .withOverrideBlocks(new CheckOutBlockOverrides()
+          .withPatronBlockOverride(new JsonObject())
+          .withComment(TEST_COMMENT)
+          .create()),
+      okapiHeaders).getJson();
+
+    item = itemsClient.get(item);
+    assertThat(item, hasItemStatus(CHECKED_OUT));
+    assertThat(loan.getString("actionComment"), is(TEST_COMMENT));
+    assertThat(loan.getString("action"), is(CHECKED_OUT_THROUGH_OVERRIDE));
   }
 
   private IndividualResource prepareLoanPolicyWithItemLimit(int itemLimit) {